+++ CPP_module/source/flexMatrix.h
@@ -97,7 +97,7 @@ public:
 		//clear matrix
 		//clear();
 
-		int numberListElements = indexI.size();
+		int numberListElements = (int)indexI.size();
 
 		//initialize vecvector
 		std::vector<int> emptyBucket(0, 0);
@@ -116,7 +116,7 @@ public:
 			int numElements = 0;
 
 			//go through bucket
-			for (int indexBucket = 0; indexBucket < buckets[indexRow].size(); indexBucket++)
+			for (int indexBucket = 0; indexBucket < (int)buckets[indexRow].size(); indexBucket++)
 			{
 				int tmpIndex = buckets[indexRow][indexBucket];
